refactor(EventStream): rename to Timeline
