+++ demo/nerdlets/nr1-community-demo-nerdlet/index.js
@@ -8,7 +8,7 @@ const componentList = [
   // { name: 'AccountDropdown', category: 'Components' },
   { name: 'DetailPanel', category: 'Components' },
   { name: 'EmptyState', category: 'Components' },
-  { name: 'EventStream', category: 'Components' },
+  { name: 'Timeline', category: 'Components' },
   { name: 'NerdGraphError', category: 'Components' },
   { name: 'timeRangeToNrql', category: 'Utilities' }
 ];
