+++ demo/nerdlets/nr1-community-demo-nerdlet/pages/index.js
@@ -1,6 +1,6 @@
 export { default as AccountDropdown } from './AccountDropdown';
 export { default as DetailPanel } from './DetailPanel';
 export { default as EmptyState } from './EmptyState';
-export { default as EventStream } from './EventStream';
+export { default as Timeline } from './Timeline';
 export { default as NerdGraphError } from './NerdGraphError';
 export { default as timeRangeToNrql } from './timeRangeToNrql';
