+++ demo/nerdlets/nr1-community-demo-nerdlet/pages/styles.scss
@@ -1,1 +1,1 @@
-@import "./EventStream/styles.scss";
+@import "./Timeline/styles.scss";
