+++ src/components/Timeline/styles.scss
