+++ src/components/index.js
@@ -1,6 +1,6 @@
 export { AccountDropdown } from './AccountDropdown';
 export { DetailPanel } from './DetailPanel';
 export { EmptyState } from './EmptyState';
-export { EventStream } from './EventStream';
+export { Timeline } from './Timeline';
 export { Funnel } from './Funnel';
 export { NerdGraphError } from './NerdGraphError';
